# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ikigai Compass API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /growth/plans:
    get:
      operationId: listGrowthPlans
      tags: [growth]
      summary: List growth plans
      responses:
        "200":
          description: List of growth plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GrowthPlan"
    post:
      operationId: createGrowthPlan
      tags: [growth]
      summary: Create growth plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGrowthPlanBody"
      responses:
        "201":
          description: Created growth plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrowthPlan"
  /growth/plans/{id}:
    get:
      operationId: getGrowthPlan
      tags: [growth]
      summary: Get growth plan with milestones
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Growth plan with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrowthPlanWithDetails"
    put:
      operationId: updateGrowthPlan
      tags: [growth]
      summary: Update growth plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGrowthPlanBody"
      responses:
        "200":
          description: Updated growth plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrowthPlan"
  /growth/milestones:
    get:
      operationId: listMilestones
      tags: [growth]
      summary: List milestones
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [growth]
      summary: Create milestone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneBody"
      responses:
        "201":
          description: Created milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /growth/milestones/{id}:
    put:
      operationId: updateMilestone
      tags: [growth]
      summary: Update milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMilestoneBody"
      responses:
        "200":
          description: Updated milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [growth]
      summary: Delete milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /growth/actions:
    get:
      operationId: listTaskActions
      tags: [growth]
      summary: List weekly task actions
      responses:
        "200":
          description: List of task actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskAction"
    post:
      operationId: createTaskAction
      tags: [growth]
      summary: Create task action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskActionBody"
      responses:
        "201":
          description: Created task action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAction"
  /growth/actions/{id}:
    put:
      operationId: updateTaskAction
      tags: [growth]
      summary: Update task action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskActionBody"
      responses:
        "200":
          description: Updated task action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAction"
  /growth/habits:
    get:
      operationId: listHabits
      tags: [growth]
      summary: List habits
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [growth]
      summary: Create habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHabitBody"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /growth/habits/{id}:
    put:
      operationId: updateHabit
      tags: [growth]
      summary: Update habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHabitBody"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [growth]
      summary: Delete habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /growth/plans/{id}/summary:
    get:
      tags: [Growth]
      summary: Get an AI-generated period summary for a growth plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Period summary for the plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeriodSummaryResponse"
        "404":
          description: Plan not found
...
components:
  schemas:
...
    GrowthPlan:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        hypothesisId:
          type: integer
        title:
          type: string
        longTermDirection:
          type: string
        ninetyDaySummary:
          type: string
        status:
          type: string
          enum: [active, paused, completed, archived]
        startDate:
          type: string
          format: date
        targetEndDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - longTermDirection
        - status
        - createdAt
        - updatedAt
    GrowthPlanWithDetails:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        hypothesisId:
          type: integer
        title:
          type: string
        longTermDirection:
          type: string
        ninetyDaySummary:
          type: string
        status:
          type: string
          enum: [active, paused, completed, archived]
        startDate:
          type: string
          format: date
        targetEndDate:
          type: string
          format: date
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/TaskAction"
        habits:
          type: array
          items:
            $ref: "#/components/schemas/Habit"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - longTermDirection
        - status
        - milestones
        - actions
        - habits
        - createdAt
        - updatedAt
    CreateGrowthPlanBody:
      type: object
      properties:
        hypothesisId:
          type: integer
        title:
          type: string
        longTermDirection:
          type: string
        ninetyDaySummary:
          type: string
        startDate:
          type: string
          format: date
        targetEndDate:
          type: string
          format: date
      required:
        - title
        - longTermDirection
    UpdateGrowthPlanBody:
      type: object
      properties:
        title:
          type: string
        longTermDirection:
          type: string
        ninetyDaySummary:
          type: string
        status:
          type: string
          enum: [active, paused, completed, archived]
        targetEndDate:
          type: string
          format: date
    Milestone:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        title:
          type: string
        description:
          type: string
        targetDate:
          type: string
          format: date
        completedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        horizon:
          type: string
          enum: [monthly, quarterly, annual]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - planId
        - title
        - status
        - horizon
        - createdAt
        - updatedAt
    CreateMilestoneBody:
      type: object
      properties:
        planId:
          type: integer
        title:
          type: string
        description:
          type: string
        targetDate:
          type: string
          format: date
        horizon:
          type: string
          enum: [monthly, quarterly, annual]
      required:
        - planId
        - title
        - horizon
    UpdateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        targetDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
    TaskAction:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        milestoneId:
          type: integer
        title:
          type: string
        description:
          type: string
        weekOf:
          type: string
          format: date
        status:
          type: string
          enum: [pending, completed, skipped]
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - planId
        - title
        - status
        - createdAt
    CreateTaskActionBody:
      type: object
      properties:
        planId:
          type: integer
        milestoneId:
          type: integer
        title:
          type: string
        description:
          type: string
        weekOf:
          type: string
          format: date
      required:
        - planId
        - title
    UpdateTaskActionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, completed, skipped]
        weekOf:
          type: string
          format: date
    Habit:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        title:
          type: string
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekdays, weekly]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - planId
        - title
        - frequency
        - isActive
        - createdAt
    CreateHabitBody:
      type: object
      properties:
        planId:
          type: integer
        title:
          type: string
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekdays, weekly]
      required:
        - planId
        - title
        - frequency
    UpdateHabitBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekdays, weekly]
        isActive:
          type: boolean
...
    PeriodSummaryResponse:
      type: object
      properties:
        headline:
          type: string
        wins:
          type: array
          items:
            type: string
        insights:
          type: array
          items:
            type: string
        nextFocus:
          type: array
          items:
            type: string
        momentumScore:
          type: number
        isAiGenerated:
          type: boolean
        planId:
          type: integer
        periodDays:
          type: integer
      required:
        - headline
        - wins
        - insights
        - nextFocus
        - momentumScore
        - isAiGenerated
        - planId
        - periodDays
...